+++ annotations/nocedit.exe/pseudocode/src/wincore/wddvmem.cpp/populateColorTable_FUN_004f4140.cpp
@@ -18,14 +18,13 @@ void __cdecl wincore_wddvmem_cpp_populateColorTable_FUN_004f4140(void)
   iVar3 = 0;
   do {
     if (g_BitsPerPixel == 0x20) {
-      uVar2 = (uint)(byte)g_SourcePaletteData[iVar1 + 2] << (g_BlueBitPosition.bytes[0] & 0x1f) |
-              (uint)(byte)g_SourcePaletteData[iVar1] << (g_RedBitPosition.bytes[0] & 0x1f) |
-              (uint)(byte)g_SourcePaletteData[iVar1 + 1] << (g_GreenBitPosition.bytes[0] & 0x1f);
+      uVar2 = (uint)g_SourcePaletteData[iVar1 + 2] << (g_BlueBitPosition.bytes[0] & 0x1f) |
+              (uint)g_SourcePaletteData[iVar1] << (g_RedBitPosition.bytes[0] & 0x1f) |
+              (uint)g_SourcePaletteData[iVar1 + 1] << (g_GreenBitPosition.bytes[0] & 0x1f);
     }
     else {
-      uVar2 = (uint)(byte)g_SourcePaletteData[iVar1] << 0x10 |
-              (uint)(byte)g_SourcePaletteData[iVar1 + 1] << 8 |
-              (uint)(byte)g_SourcePaletteData[iVar1 + 2];
+      uVar2 = (uint)g_SourcePaletteData[iVar1] << 0x10 | (uint)g_SourcePaletteData[iVar1 + 1] << 8 |
+              (uint)g_SourcePaletteData[iVar1 + 2];
     }
     iVar4 = iVar3 + 4;
     iVar1 = iVar1 + 3;
